Modify for smell

+++ src/main/java/gameobjects/Creature.java
@@ -38,5 +38,5 @@ public abstract class Creature extends DynamicEntity {
 	}
 
 	public abstract void action (float[] a);
-    public abstract byte[] observation ();
+    public abstract float[] observation ();
 }
